+++ ExcelUnitConverter/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Added
+ - Created new function ConvToSI that will convert a value to its SI unit and output that unit if done as an array formula
+
 ## [0.1.0] - 2017-02-21
 This is the initial public release of the add-in.